The_application/requirements.txt
streamlit_lottie
streamlit
bs4
pandas
torch
torch-geometric
scikit-learn
plotly
deap
joblib
Pillow
requests
io
